Handle the disc_simu.txt
The disc_simu.txt is an alternative solution to start several simulation runs. Instead of providing a range of starting materials, it specifies individual starting materials to use. 
It should be more commonly used in the 'cont' mode to re-run several simulation runs.

I used to use the format of '1, 3, 5'. But @robert-vogel  commit provides me with a better solution. So now the format is 1\n3\n5\n.